portal/requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
psycopg2
SQLAlchemy-serializer